Cleaned up uniswap class.
